Bugfix: all digests were being calculated incorrectly...
due to an extra '/' in the retrieval URL (which retrieved a 'you must be
logged in' page instead of the file contents). This affects all digests
calculated since the bug was introduced (with updates to pyclowder2 or a
change to how clowder responds to the extra / ?). Note that the extra /
was also being sent in the metadata which shows in the GUI as a bad link
(goes to the internal error page with a 'Not found' error)

Also noticed logging was not working...